rename tests -> test to enable
